// frontend/src/services/api-types.ts
export interface paths {
  "/tasks": {
    /** List tasks */
    get: {
      parameters: {
        query?: {
          eventId?: string;
        };
      };
      responses: {
        /** @description A list of tasks */
        200: {
          content: {
            "application/json": components["schemas"]["Task"][];
          };
        };
      };
    };
  };
  "/tasks/{taskId}": {
    /** Get a task by ID */
    get: {
      parameters: {
        path: {
          taskId: string;
        };
      };
      responses: {
        /** @description Task details */
        200: {
          content: {
            "application/json": components["schemas"]["Task"];
          };
        };
        /** @description Task not found */
        404: {
          content: never;
        };
      };
    };
  };
  "/tasks/{taskId}/pick": {
    /** Pick a task (assign to self) */
    post: {
      parameters: {
        path: {
          taskId: string;
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "application/json": components["schemas"]["ActionResult"];
          };
        };
        /** @description Error (e.g. Forbidden) */
        400: {
          content: {
            "application/json": components["schemas"]["ActionResult"];
          };
        };
      };
    };
  };
  "/tasks/{taskId}/transition": {
    /** Transition task state */
    post: {
      parameters: {
        path: {
          taskId: string;
        };
      };
      requestBody: {
        content: {
          "application/json": {
            nextState: components["schemas"]["TaskState"];
          };
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "application/json": components["schemas"]["ActionResult"];
          };
        };
      };
    };
  };
  "/tasks/{taskId}/assign": {
    /** Assign or unassign a task */
    post: {
      parameters: {
        path: {
          taskId: string;
        };
      };
      requestBody: {
        content: {
          "application/json": {
            userId: string | null;
          };
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "application/json": components["schemas"]["ActionResult"];
          };
        };
      };
    };
  };
  "/events": {
    /** List events */
    get: {
      responses: {
        /** @description List of events */
        200: {
          content: {
            "application/json": components["schemas"]["Event"][];
          };
        };
      };
    };
  };
  "/events/{eventId}/members": {
    /** List members of an event */
    get: {
      parameters: {
        path: {
          eventId: string;
        };
      };
      responses: {
        /** @description List of event members with user details */
        200: {
          content: {
            "application/json": (components["schemas"]["EventMember"] & {
                user?: components["schemas"]["User"];
              })[];
          };
        };
      };
    };
  };
  "/users": {
    /** List all users */
    get: {
      responses: {
        /** @description List of users */
        200: {
          content: {
            "application/json": components["schemas"]["User"][];
          };
        };
      };
    };
  };
  "/user-types": {
    /** List user types */
    get: {
      responses: {
        /** @description List of user types */
        200: {
          content: {
            "application/json": components["schemas"]["UserType"][];
          };
        };
      };
    };
  };
  "/workflow-templates": {
    /** List workflow templates */
    get: {
      responses: {
        /** @description List of workflow templates */
        200: {
          content: {
            "application/json": components["schemas"]["WorkflowTemplate"][];
          };
        };
      };
    };
    /** Save a workflow template */
    post: {
      requestBody: {
        content: {
          "application/json": components["schemas"]["WorkflowTemplate"];
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "application/json": components["schemas"]["ActionResult"];
          };
        };
      };
    };
  };
  "/workflow-instances": {
    /** List workflow instances */
    get: {
      parameters: {
        query?: {
          eventId?: string;
        };
      };
      responses: {
        /** @description List of workflow instances */
        200: {
          content: {
            "application/json": components["schemas"]["WorkflowInstance"][];
          };
        };
      };
    };
  };
  "/workflows/instantiate": {
    /** Instantiate a workflow for an event */
    post: {
      requestBody: {
        content: {
          "application/json": {
            workflowId: string;
            eventId: string;
          };
        };
      };
      responses: {
        /** @description Success */
        200: {
          content: {
            "application/json": {
              ok?: boolean;
              error?: string;
              instanceId?: string;
            };
          };
        };
      };
    };
  };
  "/task-types": {
    /** List task types */
    get: {
      responses: {
        /** @description List of task types */
        200: {
          content: {
            "application/json": components["schemas"]["TaskType"][];
          };
        };
      };
    };
  };
  "/eligibility-mappings": {
    /** List eligibility mappings */
    get: {
      responses: {
        /** @description List of eligibility mappings */
        200: {
          content: {
            "application/json": components["schemas"]["EligibilityMapping"][];
          };
        };
      };
    };
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    ActionResult: {
      ok: boolean;
      error?: string;
    };
    /** @enum {string} */
    AccessLevel: "admin" | "regular";
    /** @enum {string} */
    TaskState: "LOCKED" | "TODO" | "ASSIGNED" | "IN_PROGRESS" | "DONE";
    UserType: {
      id: string;
      name: string;
      access_level: components["schemas"]["AccessLevel"];
      permissions: {
        view: boolean;
        take: boolean;
        move_state: boolean;
        assign: boolean;
      };
    };
    User: {
      id: string;
      name: string;
      email: string;
      usertype_id: string;
      avatar_url?: string;
    };
    TaskType: {
      id: string;
      name: string;
    };
    EligibilityMapping: {
      usertype_id: string;
      tasktype_id: string;
    };
    Event: {
      id: string;
      name: string;
      description: string;
      created_at: string;
    };
    EventMember: {
      user_id: string;
      event_id: string;
    };
    WorkflowTemplate: {
      workflow_id: string;
      name: string;
      version: number;
      nodes: components["schemas"]["WorkflowNode"][];
      edges: components["schemas"]["WorkflowEdge"][];
    };
    WorkflowNode: {
      node_id: string;
      task_type_id: string;
      label: string;
      metadata: {
        description: string;
        estimated_duration_hours: number;
      };
    };
    WorkflowEdge: {
      from_node_id: string;
      to_node_id: string;
    };
    WorkflowInstance: {
      id: string;
      workflow_id: string;
      event_id: string;
      created_at: string;
    };
    Task: {
      id: string;
      workflow_instance_id: string;
      event_id: string;
      node_id: string;
      tasktype_id: string;
      label: string;
      description: string;
      state: components["schemas"]["TaskState"];
      assignee_id?: string | null;
      parent_ids: string[];
      child_ids: string[];
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}